Upgrade to version 2.2
git-svn-id: https://plugins.svn.wordpress.org/extrawatch-pro/trunk@786374 b8457f37-d9ea-0310-8a92-e5e31aec5664

+++ components/com_extrawatch/src/env/wordpress/class.extrawatch.setup.wordpress.php
@@ -4,28 +4,28 @@
  * @file
  * ExtraWatch - A real-time ajax monitor and live stats
  * @package ExtraWatch
- * @version 2.0
- * @revision 932
+ * @version 2.2
+ * @revision 1204
  * @license http://www.gnu.org/licenses/gpl-3.0.txt     GNU General Public License v3
  * @copyright (C) 2013 by CodeGravity.com - All rights reserved!
  * @website http://www.extrawatch.com
  */
 
-/** ensure this file is being included by a parent file */
-if (!defined('_JEXEC') && !defined('_VALID_MOS'))  {
-  die('Restricted access');
-}
-
-class ExtraWatchSetupWordpress implements ExtraWatchSetup
+defined('_JEXEC') or die('Restricted access');
+class ExtraWatchEnvSetupWordpress implements ExtraWatchEnvSetup
 {
+
+function __construct() {
+die("Setup construct");
+}
 
   function install()
   {
     error_reporting(E_ALL);
     $env = ExtraWatchEnvFactory::getEnvironment();
     $database = $env->getDatabase();
-    ExtraWatchSetupWordpress::create_tables($database);
-    ExtraWatchSetupWordpress::install_geolocation($database);
+    ExtraWatchEnvSetupWordpress::create_tables($database);
+    ExtraWatchEnvSetupWordpress::install_geolocation($database);
 
   }
 
@@ -33,7 +33,7 @@ class ExtraWatchSetupWordpress implements ExtraWatchSetup
   {
     $env = ExtraWatchEnvFactory::getEnvironment();
     $database = $env->getDatabase();
-    ExtraWatchSetupWordpress::drop_tables($database);
+    ExtraWatchEnvSetupWordpress::drop_tables($database);
   }
 
   function activate()
@@ -42,7 +42,7 @@ class ExtraWatchSetupWordpress implements ExtraWatchSetup
 
   static function create_tables($database)
   {
-    $sqlFile = JPATH_BASE2 . DS . "administrator" . DS . "components" . DS . "com_extrawatch" . DS . "sql" . DS . "install.mysql.utf8.sql";
+    $sqlFile = JPATH_BASE2 . DS . "components" . DS . "com_extrawatch" . DS . "sql" . DS . "install.mysql.utf8.sql";
     $sql = file_get_contents($sqlFile);
     $sql = $database->replaceDbPrefix($sql);
     $sqlSplitted = @split(";", $sql);
@@ -89,6 +89,22 @@ class ExtraWatchSetupWordpress implements ExtraWatchSetup
   static function drop_tables($database)
   {
 
+      try {
+          $database->setQuery("SELECT `value` as keepData from #__extrawatch_config where `name` = 'EXTRAWATCH_UNINSTALL_KEEP_DATA' limit 1 ");
+          $keepData = @$database->loadResult();
+      } catch (Exception $e) {
+          // suppress
+      }
+
+    if (@$keepData) {
+        echo("<b>Not deleting ExtraWatch database tables, because you set KEEP_DATA in Settings to true. <br/>Do not forget to delete these tables later, or install new version of ExtraWatch</b> <br/><br/>");
+        return;
+    }
+
+    $query = "DROP TABLE #__extrawatch_uri";
+    $database->executeQuery(trim($database->replaceDbPrefix($query)));
+    $database->query();
+
     $query = "DROP TABLE #__extrawatch";
     $database->setQuery(trim($database->replaceDbPrefix($query)));
     $database->query();
@@ -104,17 +120,15 @@ class ExtraWatchSetupWordpress implements ExtraWatchSetup
     $database->executeQuery(trim($database->replaceDbPrefix($query)));
     $database->query();
 
-    $query = "DROP TABLE #__extrawatch_ip2c";
+/* removed
+ *     $query = "DROP TABLE #__extrawatch_ip2c";
     $database->executeQuery(trim($database->replaceDbPrefix($query)));
-    $database->query();
+    $database->query();*/
 
     $query = "DROP TABLE #__extrawatch_cc2c";
     $database->executeQuery(trim($database->replaceDbPrefix($query)));
     $database->query();
 
-    $query = "DROP TABLE #__extrawatch_uri";
-    $database->executeQuery(trim($database->replaceDbPrefix($query)));
-    $database->query();
 
     $query = "DROP TABLE #__extrawatch_cache";
     $database->executeQuery(trim($database->replaceDbPrefix($query)));
@@ -162,6 +176,10 @@ class ExtraWatchSetupWordpress implements ExtraWatchSetup
     $database->query();
 
     $query = "DROP TABLE #__extrawatch_uri2keyphrase_pos";
+    $database->executeQuery(trim($database->replaceDbPrefix($query)));
+    $database->query();
+
+    $query = "DROP TABLE #__extrawatch_sql_scripts";
     $database->executeQuery(trim($database->replaceDbPrefix($query)));
     $database->query();
